change readmeteodata function for glaciers
readmeteo functions normally downscales meteo input with worldclim if meteomapsscale=False. However, even though meteo input maps need to be downscaled, the glacier input data is made available always at resolution of simulations and not at a coarser resolution. Therefore, it does not have to be downscaled, but the global input data needs to be cut to the simulation extent. This is done using the information of the ldd.